# api-gateway-spring-server/src/main/java/io/swagger/api/UserApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.62).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.User;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2024-10-01T15:58:00.855480115Z[GMT]")
@Validated
public interface UserApi {

    @Operation(summary = "Удаляет профиль авторизованного пользователя", description = "Удаляет профиль авторизованного пользователя", security = {
        @SecurityRequirement(name = "ApiTokenAuth")    }, tags={ "user" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "204", description = "Профиль пользователя успешно удалён"),
        
        @ApiResponse(responseCode = "401", description = "Несанкционированный доступ"),
        
        @ApiResponse(responseCode = "404", description = "Пользователь не найден"),
        
        @ApiResponse(responseCode = "200", description = "Ошибка сервера") })
    @RequestMapping(value = "/user",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteUser();


    @Operation(summary = "Получает данные профиля авторизованного пользователя", description = "Получает данные профиля авторизованного пользователя", security = {
        @SecurityRequirement(name = "ApiTokenAuth")    }, tags={ "user" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))),
        
        @ApiResponse(responseCode = "401", description = "Несанкционированный доступ"),
        
        @ApiResponse(responseCode = "404", description = "Пользователь не найден"),
        
        @ApiResponse(responseCode = "200", description = "Ошибка сервера") })
    @RequestMapping(value = "/user",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<User> getUser();


    @Operation(summary = "Обновляет данные авторизованного пользователя", description = "Обновляет данные авторизованного пользователя", security = {
        @SecurityRequirement(name = "ApiTokenAuth")    }, tags={ "user" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Данные пользователя успешно обновлены"),
        
        @ApiResponse(responseCode = "400", description = "Некорректные данные в запросе"),
        
        @ApiResponse(responseCode = "401", description = "Несанкционированный доступ"),
        
        @ApiResponse(responseCode = "404", description = "Пользователь не найден"),
        
        @ApiResponse(responseCode = "200", description = "Ошибка сервера") })
    @RequestMapping(value = "/user",
        consumes = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<Void> updateUser(@Parameter(in = ParameterIn.DEFAULT, description = "Данные пользователя", required=true, schema=@Schema()) @Valid @RequestBody User body
);

}