# openapi3_1__api_gateway.yaml
openapi: 3.1.0
info:
  title: Symcon smart-home — API Gateway
  description: >-
    API шлюза сервиса управления устройствами умного дома
    для взаимодействия с клиентскими приложениями.
  contact:
    name: Alexey Kuznetsov
    url: https://github.com/alexeycoder
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
servers:
  - url: https://localhost:8080/v2
    description: Test server
  - url: https://symcon.ru/v2
    description: Production server
...
paths:
  /users:

    post:
      tags:
        - user
      summary: Регистрирует нового пользователя
      description: >-
        Регистрирует нового пользователя в сервисе
        и возвращает токен авторизации
      operationId: createUser
      requestBody:
        description: Данные нового пользователя
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: Пользователь успешно создан и зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    format: byte
                    description: Токен авторизации созданного пользователя
          links:
            GetAuthorizedUser: # <---- наименование для ссылки
              operationId: getUser
              # or
              # operationRef: '#/paths/~1user~1/get'
              parameters:
                Api-Token: "$response.body#/token"
                in: header
              description: >-
                Для справки:
                Значение `token` в теле ответа используется в параметрe
                `Api-Token` заголовка в методах, требующих аутентификации
                и авторизации пользователя, например в `GET /user`.
        '400':
          description: Некорректные данные в запросе
        '409':
          description: Пользователь с таким username уже существует
        default:
          description: Ошибка сервера
  /user:

    get:
      tags:
        - user
      summary: Получает данные профиля авторизованного пользователя
      description: Получает данные профиля авторизованного пользователя
      operationId: getUser
      security:
        - ApiTokenAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Несанкционированный доступ
        '404':
          description: Пользователь не найден
        default:
          description: Ошибка сервера

    put:
      tags:
        - user
      summary: Обновляет данные авторизованного пользователя
      description: Обновляет данные авторизованного пользователя
      operationId: updateUser
      security:
        - ApiTokenAuth: []
      requestBody:
        description: Данные пользователя
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: Данные пользователя успешно обновлены
        '400':
          description: Некорректные данные в запросе
        '401':
          description: Несанкционированный доступ
        '404':
          description: Пользователь не найден
        default:
          description: Ошибка сервера

    delete:
      tags:
        - user
      summary: Удаляет профиль авторизованного пользователя
      description: Удаляет профиль авторизованного пользователя
      operationId: deleteUser
      security:
        - ApiTokenAuth: []
      responses:
        '204':
          description: Профиль пользователя успешно удалён
        '401':
          description: Несанкционированный доступ
        '404':
          description: Пользователь не найден
        default:
          description: Ошибка сервера
...
components:
  securitySchemes:
    ApiTokenAuth:
      type: apiKey
      in: header
      name: Api-Token
  schemas:
    User:
      required:
        - username
        - password
        - name
      type: object
      properties:
        username:
          type: string
          format: email
          examples:
            - johnsmith@yandex.ru
          description: Адрес эл. почты пользователя в качестве имени для входа
        password:
          type: string
          format: binary
          description: Хеш пароля
        name:
          type: string
          examples:
            - John Smith
          description: Отображаемое имя пользователя
...